refactor: add models package
