+++ models/Account.go
@@ -1,0 +1,14 @@
+package models
+
+import (
+	"time"
+)
+
+type Account struct {
+	ID        uint `gorm:"primary_key"`
+	Account   string
+	Password  string
+	Ip        string
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
